use StrictPrimitiveConverter directly
Adding the StrictPrimitiveConveter to the JsonSerializer and then
removing it was error prone (it wasn't removed in all cases already) and
unperformant. This commit switches to calling an instance of the inner
converter directly so that the original serializer can be used
unmodified while still benefitting from the strict converter.